Get less confused about what the service ID refers to.
In order to reflect what the API service is actually doing, it's an identifier for the IAPIMock provider, not an individual enpdoint, since an individual endpoint can't be reliably identified across tenants.  This also changes the URIs so that the service ID comes first, before the region, since region names only really exist *within the context* of a particular IAPIMock plugin.