add session_store to resource_for_region
This way, plugins can store all their relevant state on the session.

Start storing state that way in nova.

Note that you still have to take care to separate any data by region (test for
nova that makes sure of this for that particular case is included) because one
tenant in Identity == one Session object.