Add a Quick start section to the README
I can never remember what value I'm supposed to use for the endpoint when I want to use Mimic. Having a curl and OpenStack CLI example shows the proper value for OS_AUTH_URL settings and how to see the service catalog to get a list of other product-specific endpoints.