run in develop mode on travis
Unfortunately, due to the way that trial sets up sys.path, we're
locating the working-directory mimic as the thing to import.  So we are
not actually testing the installation procedure fully as it is, but
we're installing a mirrored copy of mimic, and telling coverage to
report coverage on the local copy.

As long as we're doing that, let's avoid doing the mirrored installation
and at least fully test the development-environment case.

Later on we should probably move to a temporary working directory before
running trial or something, so as to avoid the current dir and properly
test the full installation flow.