rewrite test-check API around check types
This change allows us to store and recall metric values per monitoring
zone in the `test-check` API. This is a significant improvement over the
initial implementation, which invented a monitoring zone for each
response. It didn't play nicely with Intelligence, which
cross-references into the monitoring zone API for labeling widgets and
things. The new implementation is much better.