feat(MaasApi): filter overview results by entity ID
This change fixes an integration bug that has been worked around
upstream in Intelligence. In this case, Intelligence sometimes requests
a single entity's overview as a performance optimization, and the
backend is assumed to return overview data for exactly one entity. Mimic
did not do anything with the `entityId` query string parameter, and
incorrect results were returned. This change adds support for the
parameter in Mimic.

Usage of the overview call, including the `entityId` query string
parameter, is described in [section 5.18.1](http://goo.gl/c6dsYu)
of the Rackspace Cloud Monitoring Developer Guide.